Version 1.0
xds_obj - An object that gets returned by xds_setup is called "an xds model object" in documentation, and in the code it is always xds_obj
Two new functions classes were added to all the components:
skill_set_*  is a description of what each module can and can not do.  This is evolving as we speak.
check_* does two things:
Right before running the model, it sets the values of terms with static values that might have changed and that trigger a need to update some downstream terms
It checks the skill set to make sure various components of the model are all internally compatible.
XH_obj - The XH component sets up something called "an XH model object" that is attached as xds_obj$XH_obj
The H component is no longer set up separately. Instead, all its objects are attached to XH object
During setup, demographic functions are set up that have no effect on the model. Advanced setup is handled by ramp.demog
Each component should either provide a method for computing terms mda and msat or report that those terms are not part of the module's skill set
MY_obj - The MY component sets up something called "an XH model object" that is attached as xds_obj$MY_obj
The protocol for setting up time-varying parameters was simplified substantially.  In a nutshell, every bionomic parameter that has a forced baseline and that can also be modified by control gets its own object, attached to MY_obj and it creates three values, the baseline, the effect size, and the value.  A function BaselineBionomics sets the baseline, and after computing effect sizes using the baseline values ModifiedBionomics sets the value of the parameter as the product of the baseline and the effect sizes.
The K_matrix respects this convention, but it has its own set of setup functions. At present, ramp.xds does not support dynamic changes in the K_matrix.  This is a planned enhancement to ramp.forcing
The eip respects this convention, but if the EIP varies with time, the object needs to include a function dF_eip
L_obj - The L component sets up something called "an XH model object" that is attached as xds_obj$MY_obj
The model objects now internalize storage of the variables indices as $ix and the initial values as $inits which makes them easier to get back.
ML-interface - the patch-habitat interface for emergence and egg laying is attached as xds_obj$ML_interface that holds things like the membership matrix, habitat search weights, and model objects that compute various terms. (It is the interface between the M and L components)
XY-interface - the blood-feeding interface is now attached as xds_obj$XY_interface that holds things like the TimeSpent matrix and objects that compute the availability of other resources. (It is the interface between the X and Y components).
There is now a formal definition of a junction that is SimBA jargon.  A junction sets up an object that does nothing but that gets used by something else. These are called by xds_compute_terms and they are:
Forcing handles variables like weather and hydrology. Functions to build non-trivial modules are found in ramp.forcing
VectorControl handles vector control. Functions to build non-trivial modules are found in ramp.control
Health handles mass drug administration and mass treatment. Functions to build non-trivial modules are found in ramp.control
xds_obj$terms - any quantity that is part of one of the interfaces that could be dynamic gets its own model object. These are attached to the appropriate interface. All of the variables that get computed by various model objects get attached to xds_obj$terms . Since these are well-defined and constrained, they are called ports (Note that the junctions set up their own sets of ports ). There's a fuzzy line between a port and junction that gets begged by Sugar but I'm gonna call it a port for now.
Habitat dynamics
Travel
Visitors
Other blood hosts
Traps
Sugar
The algorithm that does parsing was dramatically simplified and discombobulated.
All pre-derivative computation gets done by xds_compute_terms  that dispatches on xds_obj$frame in the same way as the functions that compute the derivatives. This makes it easy to pass the state variables vector and the time and recompute the state of the system at any of the time points where solutions were computed.  This was finally an elegant solution to something that had given me a great deal of trouble.
I wrote the function that computes and stores the values of the baseline and modified bionomic parameters, but I haven't called it anywhere yet.
I redefined the way steady states get computed, but I moved a lot of the qualitative analysis functions to ramp.work
I've changed the function names for consistency
make_* returns an object of some sort
setup_*  sets up something on xds_obj often calling make_*
change_* changes parameters or initial values or objects
get_* pulls something
F_*  accepts terms and computes terms