Modified steady state functions: Version 1.0.1
It was cumbersome to call the steady state functions by passing the objects, so the functions were updated to call xds_obj and the species index.

For modules that define both a system of differential equations and a discrete time system, the setup functions assign a second class to dispatch. For example, basic M defines two functions: steady_state_M.basic_M_ode and steady_state_M.basic_M_dts.